[#75] refactor: viewer context, port 위치 변경
