+++ live-chat-server/src/main/java/org/giglab/live/application/port/persistence/ViewerSessionPort.java
@@ -1,6 +1,7 @@
 package org.giglab.live.application.port.persistence;
 
 import java.util.Optional;
+import org.giglab.live.application.dto.viewer.ViewerContext;
 
 public interface ViewerSessionPort {
