+++ live-chat-server/src/main/java/org/giglab/live/global/exception/DomainErrorCode.java
@@ -1,8 +1,12 @@
 package org.giglab.live.global.exception;
 
+import org.springframework.http.HttpStatus;
+
 public interface DomainErrorCode {
 
   String getCode();
 
   String getMessage();
+
+  HttpStatus getHttpStatus();
 }
