+++ live-chat-server/src/main/java/org/giglab/live/infrastructure/redis/ViewerRedisRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.giglab.live.application.port.persistence.ViewerContext;
+import org.giglab.live.application.dto.viewer.ViewerContext;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.redis.core.Cursor;
 import org.springframework.data.redis.core.RedisCallback;
